backend/requirements.txt
# --- API Server ---
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# --- Database ---
sqlmodel>=0.0.16

# --- AI & LangChain Ecosystem (Modern Stack) ---
# Core libraries
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Integrations (New structure for LangChain 0.3)
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# --- Data Processing & Files ---
pandas>=2.2.0
numpy>=2.0.0
pydantic>=2.7.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1

# --- Utilities ---
# Table display support
tabulate>=0.9.0
requests>=2.32.0
aiofiles>=23.2.0

# --- Build Tools ---
setuptools>=70.0.0
wheel>=0.43.0

# THÊM DÒNG NÀY VÀO CUỐI:
psycopg2-binary>=2.9.9